+++ app/src/components/graph/WalkthroughNode.tsx
@@ -1,5 +1,5 @@
-import { GuideGraphNode } from "./GuideGraphNode";
-import type { GraphNodeData } from "./useGraphLayout";
+import type { GraphNodeData } from "@/lib/useGraphLayout";
+import { GuideGraphNode } from "@/components/graph/GuideGraphNode";
 
 // isSelected comes from WalkthroughGraph's getNodeState.
 type WalkthroughNodeData = GraphNodeData & { isSelected: boolean };
